feat: 팀 가용성 페이지 API 연동 (#139)
