+++ src/lib/queryKeys.ts
@@ -1,6 +1,7 @@
 export const queryKeys = {
   teams: ['teams'] as const,
   teamAvailability: ['teamAvailability'] as const,
+  teamRecommendTimes: ['teamRecommendTimes'] as const,
   events: ['events'] as const,
   auth: ['auth'] as const,
