+++ src/pages/TeamCalendar/components/SelectDurationCalendar.tsx
@@ -1,14 +1,13 @@
-import { DayPicker, useDayPicker } from 'react-day-picker';
-import type { DateRange, MonthCaptionProps } from 'react-day-picker';
-import 'react-day-picker/dist/style.css';
 import { ko } from 'date-fns/locale';
 import { ChevronLeft, ChevronRight } from 'lucide-react';
-import Button from '@/components/atoms/Button';
+import type { DateRange, MonthCaptionProps } from 'react-day-picker';
+import { DayPicker, useDayPicker } from 'react-day-picker';
+import 'react-day-picker/dist/style.css';
 
 interface Props {
   range: DateRange | undefined;
-  setRange: (range: DateRange | undefined) => void;
-  onSearch: () => void;
+  setRange?: (range: DateRange | undefined) => void;
+  onSearch?: () => void;
 }
 
 // 커스텀 캡션 컴포넌트
@@ -44,7 +43,7 @@ function CustomCaption(props: MonthCaptionProps) {
   );
 }
 
-const SelectDurationCalendar: React.FC<Props> = ({ range, setRange, onSearch }) => {
+const SelectDurationCalendar: React.FC<Props> = ({ range, setRange }) => {
   return (
     <div className="date-picker">
       {/* 캘린더 컨테이너 */}
@@ -53,8 +52,9 @@ const SelectDurationCalendar: React.FC<Props> = ({ range, setRange, onSearch }) 
           <DayPicker
             mode="range"
             selected={range}
-            onSelect={setRange}
+            onSelect={setRange ? setRange : undefined}
             locale={ko} // 한국어 요일 헤더
+            disabled={() => true} // 모든 날짜 클릭 비활성화
             components={{
               MonthCaption: CustomCaption, // 커스텀 캡션 사용
             }}
@@ -94,14 +94,14 @@ const SelectDurationCalendar: React.FC<Props> = ({ range, setRange, onSearch }) 
           </div>
 
           {/* 검색 버튼 */}
-          <Button
+          {/* <Button
             onClick={range?.from && range?.to ? onSearch : () => {}}
             text="일정 검색하기"
             fullWidth={true}
             variant={range?.from && range?.to ? 'primary' : 'secondary'}
             size="md"
             className={`${!range?.from || !range?.to ? 'cursor-not-allowed opacity-50' : ''}`}
-          />
+          /> */}
         </div>
       </div>
     </div>
