+++ src/utils/dateTimeUtils.ts
@@ -74,3 +74,19 @@ export const formatDateWithWeekday = (isoString: string): string => {
 
   return `${month}월 ${day}일 ${weekday}요일`;
 };
+
+/**
+ * Date 객체를 LocalDateTime 형식으로 변환 (Z 없이)
+ * @param date Date 객체
+ * @returns LocalDateTime 형식의 문자열 (예: "2025-11-01T00:00:00")
+ */
+export const toLocalDateTime = (date: Date): string => {
+  const year = date.getFullYear();
+  const month = String(date.getMonth() + 1).padStart(2, '0');
+  const day = String(date.getDate()).padStart(2, '0');
+  const hours = String(date.getHours()).padStart(2, '0');
+  const minutes = String(date.getMinutes()).padStart(2, '0');
+  const seconds = String(date.getSeconds()).padStart(2, '0');
+
+  return `${year}-${month}-${day}T${hours}:${minutes}:${seconds}`;
+};
